requirements.txt
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
MouseInfo==0.1.3
numpy==2.2.6
opencv-python==4.12.0.88
pillow==10.2.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.1.1
pytweening==1.2.0
requests==2.32.4
six==1.17.0
urllib3==2.5.0
SpeechRecognition
gTTS
playsound==1.2.2
pyaudio
pyttsx3
pycaw
comtypes
apscheduler
pystray
pillow
keyboard
pyinstaller
pyinstaller-hooks-contrib
pynput
